Fixed warnings

+++ NBTMeister/NBTMeister/libs/zlib-contrib/zfstream.cpp
@@ -194,7 +194,7 @@ gzfilebuf::underflow()
 
   // Attempt to fill internal buffer from gzipped file
   // (buffer must be guaranteed to exist...)
-  int bytes_read = gzread(file, buffer, buffer_size);
+  int bytes_read = gzread(file, buffer, (unsigned int)buffer_size);
   // Indicates error or EOF
   if (bytes_read <= 0)
   {
@@ -226,7 +226,7 @@ gzfilebuf::overflow(int_type c)
       this->pbump(1);
     }
     // Number of characters to write to file
-    int bytes_to_write = this->pptr() - this->pbase();
+    int bytes_to_write = (int)(this->pptr() - this->pbase());
     // Overflow doesn't fail if nothing is to be written
     if (bytes_to_write > 0)
     {
